+++ src/pl/charmas/parcelablegenerator/typeserializers/serializers/ListPrimitiveSerializer.java
@@ -16,25 +16,39 @@
 package pl.charmas.parcelablegenerator.typeserializers.serializers;
 
 import com.intellij.psi.PsiField;
+import com.intellij.psi.PsiType;
+
+import java.util.List;
+
 import pl.charmas.parcelablegenerator.typeserializers.TypeSerializer;
+import pl.charmas.parcelablegenerator.util.PsiUtils;
 
 public class ListPrimitiveSerializer implements TypeSerializer {
 
-    private final String typeName;
+    public String writeValue(String parcel, String assignTo) {
+        return parcel + ".writeList(" + assignTo + ");";
+    }
 
-    public ListPrimitiveSerializer(String typeName) {
-        this.typeName = typeName;
+    public String readValue(String genericType, String assignTo) {
+        return assignTo + " = new java.util.ArrayList<" + genericType + ">();" +
+                       "in.readList(" + assignTo + " ," + genericType + ".class.getClassLoader());";
     }
 
     @Override
     public String writeValue(PsiField field, String parcel, String flags) {
-        return parcel + ".writeList(this." + field.getName() + ");";
+        final String fieldName = field.getName();
+        return writeValue(parcel, "this." + fieldName);
     }
 
     @Override
     public String readValue(PsiField field, String parcel) {
-        return "this." + field.getName() + " = new java.util.ArrayList<" + typeName + ">();" +
-                "\n" +
-                "in.readList(this." + field.getName() + " ," + typeName + ".class.getClassLoader());";
+        final List<PsiType> resolvedGenerics = PsiUtils.getResolvedGenerics(field.getType());
+
+        // Assumes a List never has more than one generic type
+        String genericType = resolvedGenerics.get(0).getCanonicalText();
+
+        final String fieldName = field.getName();
+
+        return readValue(genericType, "this." + fieldName);
     }
 }