+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 
+##Version 0.6
+
+* CREATOR is now final - proguard will not touch it now
+* fixed parcelable class loader
+
 ##Version 0.2
 
 * added nullable primitives support
-* added primite array support
+* added primitive array support
 * added boolean sparse array support
 * added list list serialization